generalize pattern to base class
[#146663059]
