+++ lib/api/meta/relation_collections/get_for.rb
@@ -2,31 +2,9 @@
 module API
   module Meta
     module RelationCollections
-      class GetFor
-        def initialize(endpoint_klass, opts = {}, &block)
-          @endpoint_klass = endpoint_klass
-          @opts = opts
-          opts[:block] = block if block.present?
-
-          decorate
-        end
+      class GetFor < Meta::Base
 
         private
-
-        attr_reader :endpoint_klass, :opts
-
-        def decorate
-          return unless opts.present?
-
-          endpoint_klass.send(:desc, desc)
-          endpoint_klass.send(
-            :route_setting,
-            scopes: opts[:scopes]
-          ) if opts.key?(:scopes)
-
-          define_params
-          define_endpoint
-        end
 
         def define_params
           endpoint_klass.instance_exec(this_resource) do |resource_name|
@@ -65,10 +43,6 @@ module API
         def desc
           @desc ||= "Get #{opts[:relation].to_s.singularize} for a given "\
                     "#{this_resource.downcase}"
-        end
-
-        def this_resource
-          @this_resource ||= endpoint_klass.name.demodulize
         end
       end
     end
