+++ lib/api/meta/simple_reads.rb
@@ -1,0 +1,10 @@
+module API
+  module Meta
+    module SimpleReads
+      extend ActiveSupport::Autoload
+      autoload :GetById
+
+      include Meta::Delegator
+    end
+  end
+end
