+++ lib/api/meta/simple_reads/get_by_id.rb
@@ -1,0 +1,8 @@
+module API
+  module Meta
+    module SimpleReads
+      class GetById
+      end
+    end
+  end
+end
